requirements.txt
pandas==1.2.4
protobuf==3.19.1
SQLAlchemy==1.4.7
google-api-core==2.3.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.31.0
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0